move display and to_sql_string impls to separate modules for plan
