+++ core/translate/optimizer/mod.rs
@@ -50,7 +50,7 @@ pub fn optimize_plan(plan: &mut Plan, schema: &Schema) -> Result<()> {
     // Context for everything is created with values inside the plan
     tracing::debug!(
         target = "optimized_plan",
-        plan_sql = plan.to_sql_string(&crate::translate::plan::PlanContext(&[]))
+        plan_sql = plan.to_sql_string(&crate::translate::display::PlanContext(&[]))
     );
     Ok(())
 }
